refactor: clean up nfa list code
