+++ ui/src/views/explore/explore-list/explore-list-container.tsx
@@ -1,16 +1,12 @@
-import { Flex } from '@/components';
-
 import { Explore } from '../explore.context';
-import { NFAListFragment } from './nfa-list';
+import { NFAListFragment } from './nfa-list.fragment';
 import { NFASearchFragment } from './nfa-search.fragment';
 
 export const ExploreListContainer: React.FC = () => {
   return (
-    <Flex css={{ flexDirection: 'column' }}>
-      <Explore.Provider>
-        <NFASearchFragment />
-        <NFAListFragment />
-      </Explore.Provider>
-    </Flex>
+    <Explore.Provider>
+      <NFASearchFragment />
+      <NFAListFragment />
+    </Explore.Provider>
   );
 };
