+++ ui/src/views/explore/explore-list/nfa-list.fragment.tsx
@@ -1,11 +1,12 @@
 import { useQuery } from '@apollo/client';
 import { useEffect } from 'react';
 
-import { Flex, NFACard, NFACardSkeleton } from '@/components';
+import { NFACard, NFACardSkeleton } from '@/components';
 import { lastNFAsPaginatedDocument } from '@/graphclient';
 import { useWindowScrollEnd } from '@/hooks';
 
-import { Explore } from '../../explore.context';
+import { Explore } from '../explore.context';
+import { NFAListFragmentStyles as S } from './nfa-list.styles';
 
 const pageSize = 10; //Set this size to test pagination
 
@@ -61,29 +62,16 @@ export const NFAListFragment: React.FC = () => {
   if (queryError) return <div>Error</div>; //TODO handle error
 
   return (
-    <Flex
-      css={{
-        flexDirection: 'column',
-        gap: '$2',
-        my: '$6',
-        minHeight: '50vh',
-        marginBottom: '30vh', // TODO: remove this if we add page footer
-      }}
-    >
-      <Flex css={{ gap: '$6', flexWrap: 'wrap' }}>
-        {tokens.map((token) => (
-          <NFACard data={token} key={token.id} />
-        ))}
-        {isLoading && <LoadingSkeletons />}
-        {!isLoading && tokens.length === 0 && (
-          // TODO: update this after designs are done
-          <div
-            className={`relative cursor-default select-none pt-2 px-3.5 pb-4 text-slate11 text-center`}
-          >
-            Nothing found.
-          </div>
-        )}
-      </Flex>
-    </Flex>
+    <S.Container>
+      {tokens.map((token) => (
+        <NFACard data={token} key={token.id} />
+      ))}
+
+      {isLoading && <LoadingSkeletons />}
+
+      {!isLoading && tokens.length === 0 && (
+        <S.EmptyMessage>Nothing found.</S.EmptyMessage>
+      )}
+    </S.Container>
   );
 };
