--- ui/src/views/explore/explore-list/nfa-list/index.ts
@@ -1,1 +1,0 @@
-export * from './nfa-list';
