+++ ui/src/views/explore/explore.styles.ts
@@ -4,7 +4,5 @@ import { styled } from '@/theme';
 export abstract class Explore {
   static readonly Container = styled(Flex, {
     flexDirection: 'column',
-    width: '64.75rem', //TODO replace for max-width
-    margin: '0 auto',
   });
 }
